Module Pyplot.Plot

type color =

| Red

| Green

| Blue

| Rgb of {

r : float;

g : float;

b : float;

}

type shape =

| Square

| Dot

| Circle

| Triangle

type style = color * shape

type linestyle =

| Line_solid

| Line_dashed

| Line_dashdot

| Line_dotted

| Line_nothing

type line = {

linestyle : linestyle;

dot_style : shape option;

color : color;

}

type dim2

type dim3

type _ scatter_data =

| Dim2Scatter : {

xs : Matrix.t;

ys : Matrix.t;

} -> dim2 scatter_data

| Dim3Scatter : {

xs : Matrix.t;

ys : Matrix.t;

zs : Matrix.t;

} -> dim3 scatter_data

type _ axes =

| Dim2Axes : {

xaxis : string;

yaxis : string;

} -> dim2 axes

| Dim3Axes : {

xaxis : string;

yaxis : string;

zaxis : string;

} -> dim3 axes

type 'dim scatter =

| Scatter : {

data : ('dim scatter_data * style) list;

axes : 'dim axes;

title : string;

} -> 'dim scatter

type histogram =

| Histogram of {

data : Matrix.t;

options : histogram_options option;

axes : dim2 axes;

title : string;

}

and histogram_options = {

bins : bins_options option;

range : (float * float) option;

}

and bins_options =

| Bins_num of int

| Bins_edges of float list

| Bins_auto

| Bins_fd

| Bins_sturges

| Bins_doane

| Bins_scott

| Bins_stone

| Bins_rice

| Bins_sqrt

module Axis : sig ... end

Plots are called "Axes" in matplotlib. For now we provide 2d & 3d scatterplots and histograms. We abstract monadically over matplotlib's statefulness.

type 'a t

The type of layouts of plots. This specifies how a plot should be displayed. matplotlib allows to compose plots together as subplots. We provide another monadic interface to abstract over matplotlib's statefulness.

We can't use the same monad as in the Axis module because the corresponding states are handled distinctly in matplotlib.

val return : 'a -> 'a t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val run : ?⁠figsize:(float * float) -> nrows:int -> ncols:int -> 'a t -> 'a

Performs a layout in a matrix of nrowsxncols subplots.

val subplot_2d : row:int -> col:int -> (unit, dim2) Axis.t -> unit t

Add a 2d subplot at the prescribde position in the ambient subplot matrix.

val subplot_3d : row:int -> col:int -> (unit, dim3) Axis.t -> unit t

Add a 3d subplot at the prescribde position in the ambient subplot matrix.

val scatter : 'a scatter -> (unit, 'a) Axis.t

Declarative helper to produce a 2d or 3d scatterplot.

val histogram : histogram -> (unit, dim2) Axis.t

Declarative helper to produce a histogram.

val suptitle : title:string -> fontsize:int option -> unit t

Sets title for whole figure.

val savefig : filename:string -> dpi:int -> quality:int -> unit t

Saves the current layout to a file.

val show : unit -> unit t

Displays the current layout on the screen.

val init : ?⁠interpreter:string -> ?⁠version:int -> unit -> unit

Initializes python. This function has to be called first.