Module Plot.Axis

Plots are called "Axes" in matplotlib. For now we provide 2d & 3d scatterplots and histograms. We abstract monadically over matplotlib's statefulness.

type ('a, 'k) t

val return : 'a -> ('a, 'k) t

val (>>=) : ('a, 'k) t -> ('a -> ('b, 'k) t) -> ('b, 'k) t

val run : ('a, 'k) t -> Pytypes.pyobject -> 'a

val set_xlabel : string -> (unit, 'k) t

Set the x-axis label for the current plot.

val set_ylabel : string -> (unit, 'k) t

Set the y-axis label for the current plot.

val set_zlabel : string -> (unit, 'k) t

Set the z-axis label for the current plot.

val set_title : string -> (unit, 'k) t

Set the title for the current plot.

val set_ylim : ymin:float option -> ymax:float option -> (unit, 'k) t

Override the range of the y axis manually.

val histogram_1d : h:Matrix.t -> opts:histogram_options -> (unit, dim2) t

Plots a histogram. The matrix h contains columns(h) histogram. I.e. data for each histograms are stored as column vectors.

val scatter_2d : xs:Matrix.t -> ys:Matrix.t -> color:color -> shape:shape -> (unit, dim2) t

2d scatter plot. xs stores the x-components of the points and ys the y-components of the points. xs and ys should be row vectors of the same dimensions. color specifies the color used for the plot. shape specifies the shape of each point in the scatter plot.

val scatter_3d : xs:Matrix.t -> ys:Matrix.t -> zs:Matrix.t -> color:color -> shape:shape -> (unit, dim3) t

See scatter_2d.

val line_2d : xs:Matrix.t -> ys:Matrix.t -> legend:string option -> line:line option -> (unit, dim2) t

2d line plot. xs stores the x-components and ys the y-components. line is an optional style, by default a red solid line will be drawn.

val legend : unit -> (unit, 'k) t

Activates displaying legends defined up to call time for plots in current axis.